requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
email-validator==2.1.1
scikit-learn==1.4.2
pandas==2.2.2
reportlab==4.1.0
